chore(FindProjectRoot): Move func to utils pkg to avoid code duplication
